backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
langchain==0.0.345
groq==0.4.1
chromadb==0.4.18
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pymongo==4.6.0
langchain-community
email-validator
huggingface_hub==0.15.1